[CAMEL-5990]Upgrade to QuickFIX/J 1.5.3
git-svn-id: https://svn.apache.org/repos/asf/camel/branches/camel-2.10.x@1457621 13f79535-47bb-0310-9956-ffa450edef68

+++ components/camel-quickfix/src/test/java/org/apache/camel/component/quickfixj/examples/trading/TradeExecutor.java
@@ -162,9 +162,9 @@ public class TradeExecutor {
 
     private boolean isOrderExecutable(Message order, Price price) throws FieldNotFound {
         if (order.getChar(OrdType.FIELD) == OrdType.LIMIT) {
-            BigDecimal limitPrice = new BigDecimal(order.getString(Price.FIELD));
+            Double limitPrice = new Double(order.getString(Price.FIELD));
             char side = order.getChar(Side.FIELD);
-            BigDecimal thePrice = price.getValue();
+            Double thePrice = price.getValue();
 
             return (side == Side.BUY && thePrice.compareTo(limitPrice) <= 0) 
                 || ((side == Side.SELL || side == Side.SELL_SHORT) && thePrice.compareTo(limitPrice) >= 0);
